finish refactor of utils folder (Other repo and coding issues section 3)
