+++ learning_lidar/utils/misc_lidar.py
@@ -23,14 +23,12 @@
 Miscellaneous operations for lidar analysis and calculations.
 """
 
-import os, sys
 import numpy as np
-import logging
 import pandas as pd
 import matplotlib.pyplot as plt
 from matplotlib import rc
 import scipy.sparse as sprs
-from datetime import datetime, timedelta, time
+from datetime import timedelta
 
 rc('font', **{'family': 'sans-serif', 'sans-serif': ['Helvetica']})
 ## for Palatino and other serif fonts use:
@@ -202,61 +200,6 @@ class LidarProfile(object):
         return self.__getitem__[key]
 
 
-def smooth(x, window_len=11, window='hanning'):
-    """smooth the data using a window with requested size.
-
-    This method is based on the convolution of a scaled window with the signal.
-    The signal is prepared by introducing reflected copies of the signal
-    (with the window size) in both ends so that transient parts are minimized
-    in the begining and end part of the output signal.
-
-    input:
-        x: the input signal
-        window_len: the dimension of the smoothing window; should be an odd integer
-        window: the type of window from 'flat', 'hanning', 'hamming', 'bartlett', 'blackman'
-            flat window will produce a moving average smoothing.
-
-    output:
-        the smoothed signal
-
-    example:
-
-    t=linspace(-2,2,0.1)
-    x=sin(t)+randn(len(t))*0.1
-    y=smooth(x)
-
-    see also:
-
-    numpy.hanning, numpy.hamming, numpy.bartlett, numpy.blackman, numpy.convolve
-    scipy.signal.lfilter
-
-    TODO: the window parameter could be the window itself if an array instead of a string
-    NOTE: length(output) != length(input), to correct this: return y[(window_len/2-1):-(window_len/2)] instead of just y.
-    """
-
-    if x.ndim != 1:
-        raise ValueError("smooth only accepts 1 dimension arrays.")
-
-    if x.size < window_len:
-        raise ValueError("Input vector needs to be bigger than window size.")
-
-    if window_len < 3:
-        return x
-
-    if not window in ['flat', 'hanning', 'hamming', 'bartlett', 'blackman']:
-        raise ValueError("Window is on of 'flat', 'hanning', 'hamming', 'bartlett', 'blackman'")
-
-    s = np.r_[x[window_len - 1:0:-1], x, x[-2:-window_len - 1:-1]]
-    # print(len(s))
-    if window == 'flat':  # moving average
-        w = np.ones(window_len, 'd')
-    else:
-        w = eval('np.' + window + '(window_len)')
-
-    y = np.convolve(w / w.sum(), s, mode='valid')
-    return y
-
-
 def calc_tau(sigma, heights):
     """
     Calculate the the attenuated optical depth (tau is the optical depth), through a collumn of heights.
@@ -329,7 +272,7 @@ def calc_extiction_klett(S, heights, sigma_0, ind_m, k=1):
 
     dr = heights[1] - heights[0]
     exp_S = np.exp((S - S_m) / k)
-    denominator = (1 / (sigma_m) + (2 / k) * np.flip(np.cumsum(np.flip(exp_S, 0) * dr), 0))  # backward integration
+    denominator = (1 / sigma_m + (2 / k) * np.flip(np.cumsum(np.flip(exp_S, 0) * dr), 0))  # backward integration
     sigma = exp_S / (denominator + eps)
     return sigma
